Space optimization

+++ include/MagicConstants.h
@@ -7,19 +7,19 @@
  */
 
 #pragma once
-
+#include "Util.h"
 
 #define um *1
 #define mm *1000
 #define cm *1000*10
 #define m  *1000*1000
 
-constexpr size_t CELL_SIZE = 20 um;
-constexpr size_t FIBER_LEN = 3 cm / CELL_SIZE;
-constexpr size_t DEFAULT_B = 100 um;
-constexpr size_t DEFAULT_C = 100 um;
+constexpr uint CELL_SIZE = 20 um;
+constexpr uint FIBER_LEN = 3 cm / CELL_SIZE;
+constexpr uint DEFAULT_B = 100 um;
+constexpr uint DEFAULT_C = 100 um;
 // for every 50 cells, generate a random fiber
-constexpr size_t C_PER_FIBER = 50; 
+constexpr uint C_PER_FIBER = 50;
 
 #undef um
 #undef mm
